feat(rhino): Added rhino package
Added package rhino that will subscribe to an mqtt topic,
transform the data for future consumers and publish it to
a kafka topic

Type rhino is the base of this package, connecting to kafka and
creating a topic if needed. rhino also provides a handler for
the mqtt client Subscribe method that will connect these two
data streams.