feat(lion): Add PathFinder
Added interface PathFinder and type DumbPathFinder
DumbPathFinder uses bus points and the new geopoint to create a request to the local GraphHopper API
to get the path the bus should take to reach all listed points.
Path is returned as a polyline encoded string.
Added integration test to make sure it worked with local GraphHopper server